+++ internal/auth/auth_test.go
@@ -18,20 +18,52 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-package utils
+package auth
 
 import (
-	"log"
 	"net/http"
+	"testing"
 )
 
-func HandleReadiness() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Add("content-type", "text/plain; charset=utf-8")
-		w.WriteHeader(http.StatusOK)
-		_, err := w.Write([]byte(http.StatusText(http.StatusOK)))
-		if err != nil {
-			log.Fatal(err)
-		}
+func TestGetHeaderToken_MultipleValues(t *testing.T) {
+	tests := []struct {
+		name    string
+		headers http.Header
+		want    string
+		wantErr error
+	}{
+		{
+			name:    "Single value in Api-Key header",
+			headers: http.Header{"Api-Key": []string{"value1"}},
+			want:    "value1",
+		},
+		{
+			name:    "No Api-Key header",
+			headers: http.Header{},
+			wantErr: ErrNoAuthHeaderIncluded,
+		},
+		{
+			name:    "Empty Api-Key header",
+			headers: http.Header{"Api-Key": []string{""}},
+			wantErr: ErrNoAuthHeaderIncluded,
+		},
+		{
+			name:    "Empty Api-Key header with other headers",
+			headers: http.Header{"Api-Key": []string{""}, "Other-Header": []string{"value"}},
+			wantErr: ErrNoAuthHeaderIncluded,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := GetHeaderToken(tt.headers)
+			if err != tt.wantErr {
+				t.Errorf("GetHeaderToken() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			if got != tt.want {
+				t.Errorf("GetHeaderToken() got = %v, want %v", got, tt.want)
+			}
+		})
 	}
 }
