+++ packages/convert2mp3/convert2mp3.go
@@ -1,3 +1,23 @@
+// Copyright (c) 2024 Charles Ozochukwu
+
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
 package convert2mp3
 
 import (
@@ -13,17 +33,19 @@ type Response struct {
 
 func ConvertToMp3(w http.ResponseWriter, r *http.Request, cfg *config.ApiConfig) {
 	ctx := r.Context()
+	// creates a unique folder within the current directory
 	folderPath, err := utils.CreateUniqueFolder(utils.BasePath)
 	if err != nil {
 		utils.RespondWithError(w, http.StatusInternalServerError, err.Error())
 		return
 	}
+	// processes the uploaded file and converts it to mp3, then saves it to the unique folder path
 	outputfileName, err := handleRequestBodyAndConvertToMP3(r, folderPath)
 	if err != nil {
 		utils.RespondWithError(w, http.StatusInternalServerError, err.Error())
 		return
 	}
-	// upload file to cloudinary
+	// uploads the file to cloudinary to get back the direct url link
 	urlLink, err := utils.HandleFileUpload(ctx, outputfileName, cfg.CloudinaryUrl)
 	if err != nil {
 		utils.RespondWithError(w, http.StatusInternalServerError, err.Error())
@@ -35,5 +57,6 @@ func ConvertToMp3(w http.ResponseWriter, r *http.Request, cfg *config.ApiConfig)
 		utils.RespondWithError(w, http.StatusInternalServerError, err.Error())
 		return
 	}
+	//returns back the response in JSON format
 	utils.RespondWithJSON(w, http.StatusOK, Response{Response: urlLink})
 }
