+++ packages/replicate/generateimages/handlegenerations.go
@@ -1,3 +1,22 @@
+// Copyright (c) 2024 Charles Ozochukwu
+
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE
 package generateimages
 
 import (
@@ -11,15 +30,6 @@ import (
 	replicate "github.com/replicate/replicate-go"
 	"net/http"
 )
-
-//	var imagemodelInputProcessors = map[int]func(ctx context.Context, r *http.Request, token string, modelIndex int) (replicate.PredictionInput, error){
-//		0: processLowImageGenerationInput,
-//		1: processHighImageGenerationInput,
-//		2: processHighImageGenerationInput,
-//		3: processLowImageGenerationInput,
-//		4: processHighImageGenerationInput,
-//		// Add other model processors similarly...
-//	}
 
 func processImageModelInput(imageModel *rep.ReplicateModel, ctx context.Context, r *http.Request, modelIndex int, cfg *config.ApiConfig) (replicate.PredictionInput, error) {
 	if imageModel.Category == "Low" {
